Fixed up the inlet... I was doing it backwards, and didn't realize it right away.
Now it's a bit bigger.  I'm debating if it needs a bigger slope on the bottom.

Also the bare inlet now has a hole for the marble to come out.  I didn't think it should have one, but with the sloped base I thought it wise to put the hole in.  I'll upload other options for the module later.