Longer pegs, modularizing.
I'm working up a single-shot module - a lifter that can recirculate OR pass marbles on - so you can easily print a single panel, or make 100 and they'll still link up.

This will be a requirement for lifters going forwards - that they can be modified to accept marbles on the left, and deposit them on the right at the same height.

Remaining work is to make the spiral drop go straight down - which will let you swap from return to pass on easily.

Bonus points: a switch so you can choose recirculate/pass on at will.

Paul