Deleted all the STLs!  Cleaning up the lifters.
Also working on attaching the 12x12 pegboards to a bigger pegboard:
The rear ball return and the legs now extend exactly 1" behind the board.
Next, plan to make a mount to attach the handle to another pegboard.

idea 1: use regular pegboard hooks to hang an adapter, which pokes through one or
two of the handle mounting holes.

Paul